fix(TCOMP-911): move Suggestion classes to suggestion package
